+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.4.17 Release:
+## New Content:
+- Druid: Wild Shape (Beta)
+- Circle of the Moon: Circle Forms and Primal Strike
+- Item: Insignia of Claws
+## Bug Fixes:
+- Steel Defender
+## Update Notes:
+- Wild Shape may still need some more adjustments / fixes.  Make a backup of your actor before use and report any bugs.
+  
 # 0.4.16 Release:
 ## New Content:
 - Battle Smith Artificer: Steel Defender and Arcane Jolt (Thanks Autumn)